requirements.txt
opencv-python
numpy
pandas
Pillow
openpyxl
tkinter